structure functions in a utils file
